ap/pv added, version 1.6.20.
PV ("Pipe Viewer") is a tool for monitoring the progress of data
through a pipeline.  It can be inserted into any normal pipeline
between two processes to give a visual indication of how quickly
data is passing through, how long it has taken, how near to
completion it is, and an estimate of how long it will be until
completion.

URL: http://www.ivarch.com/programs/pv.shtml
Signed-off-by: Michal Cieslakiewicz <michal.cieslakiewicz@wp.pl>