ap/podman-catatonit added, version 0.1.7.
A container init that is so simple it's effectively brain-dead. This
is a rewrite of initrs in C - which, in turn, is a reimplementation
of other container inits like tini and dumb-init.
This version integrates itself with podman by putting its binary in
podman's default helper search path.

URL: https://github.com/openSUSE/catatonit
Signed-off-by: Michal Cieslakiewicz <michal.cieslakiewicz@wp.pl>