xap/gsmartcontrol added, version 1.1.4.
GSmartControl is a graphical user interface for smartctl (from
smartmontools package), which is a tool for querying and controlling
SMART (Self-Monitoring, Analysis, and Reporting Technology) data on
modern hard disk and solid-state drives. It allows you to inspect
the drive's SMART data to determine its health, as well as run
various tests on it. This tool automatically reports and highlights
any anomalies.

URL: https://github.com/ashaduri/gsmartcontrol

Signed-off-by: Michal Cieslakiewicz <michal.cieslakiewicz@wp.pl>