ap/pla-util added, version 2.1.3.
The purpose of this program is to allow Linux users to manage
HomePlug AV2 compliant power line adapters that use a Broadcom
system-on-a-chip (SoC) such as the BCM60500 or BCM60333.

URL: https://github.com/serock/pla-util

Note: cross-compiling for different architectures is not
supported. For x86-64 default is "-mtune=generic" so it should
be safe for most CPUs.