n/chrony added, version 4.6.1.
chrony is a versatile implementation of the Network Time Protocol.
It can synchronise the system clock with NTP servers, reference clocks
(e.g. GPS receiver), and manual input using wristwatch and keyboard.
It can also operate as an NTPv4 (RFC 5905) server and peer to provide
a time service to other computers in the network.

URL: https://chrony-project.org/

Note: Network Manager integration is skipped for now.